delete doesn't block exceptions propagation even for AttributeError and UnmappedInstanceError

+++ sprox/mg/provider.py
@@ -406,10 +406,8 @@ class MingProvider(IProvider):
     def delete(self, entity, params):
         """Delete an entry of typeentity which matches the params."""
         obj = self.get_obj(entity, params)
-        try:
+        if obj is not None:
             obj.delete()
-        except AttributeError:
-            pass  # It is idempotent to delete something that does not exists
         return obj
 
     def _modify_params_for_related_searches(self, entity, params, view_names=None, substrings=()):
